Remove user-space mounting tool
There is no need in a separate tool to mount a fuse directory. Linux needs one
because mount() syscall is preveleged and there should be a binary with
user bit set. But on MacOSX mount() is unprevileged - just mounting with syscall
works perfectly fine (and much faster, simpler and easier to support).