Remove KILL_ON_UNMOUNT option.
According to macfuse documentation this is a hack for MacOSX 10.4:

On Mac OS X 10.4.x, a user-space file system may not exit upon unmount. This is because it may be blocked in one or more operations whose cancellation is not supported by the operating system. The kill_on_unmount option can be used as a workaround to make the kernel send a SIGKILL to the file system daemon /after/ the unmount call has been processed. Although I have not tested exhaustively, this should not be necessary on Mac OS X 10.5.x.