Add fioc example that shows how ioctl() works
Linux allows to call ioctl() on regular files, it is up to filesystem to decide
on what types of files this call is allowed.

In XNU kernel checks if the file is a special device. If it is not ithen kernel returns
ENOTTY error and vfs is never called.

Fix the example by marking filetype as block device file.